login rest

+++ app/src/main/java/com/example/bitcard/network/daos/responses/GetUserResponse.kt
@@ -3,7 +3,7 @@ package com.example.bitcard.network.daos.responses
 import com.example.bitcard.network.daos.requests.UserModel
 
 data class GetUserResponse(
-    val status_code : Long,
+    val status_code : Int,
     val description: String,
     val data: UserModel
 )
